refactor: rename era file to era types before introducing era file reader/writer
