+++ crates/era/src/lib.rs
@@ -7,4 +7,4 @@
 
 pub mod beacon_types;
 pub mod e2s_types;
-pub mod era_file;
+pub mod era_types;
